% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state() plot the incidents in a state for a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a string or integer to indicate a state}

\item{year}{a string or integer to indicate a year}
}
\value{
return a map indicates total number of accidents from the input state number and input year
}
\description{
This function read a number of a state and a year, and then plot a map of 
the accidents happened in that state in that year
}
\details{
An error will occur if the state number is wroing
}
\examples{
\dontrun{
   fars_map_state(45, 2013)
}

}

