% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read() read and convert to dataset}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a charater string indicates name of the input file}
}
\value{
this function returns a data frame which is converted from input file
}
\description{
This function check if the file exists and then create a data frame
}
\details{
An error will be generated if the file does not exist.
}
\examples{
\dontrun{
   fars_read(inputfile)
}

}

