% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years() generate a list of data frame}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years, can be string or integer}
}
\value{
a list of data frames consists of MONTH and year columns
}
\description{
This function is to read a vector of years, and load all the files "accident_year.csv.bz2", where 
years is the input year, and then return a list of data frame that only include the MONTH and year
which is from input
}
\details{
An error will occure if the input year does not match the years in the file name
}
\examples{
\dontrun{
   fars_read_years(c(2013:2015))
}

}

