% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years() generate a data frame to summary month incidence}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years, can be string or integer}
}
\value{
a data frame to summary the incidence for each month
}
\description{
This function reads a vector of year and then first generate data frame with column MONTH and year
through function fars_read_years(), and then return a data frame shows the incidence for each month
}
\examples{
\dontrun{
   fars_read_years(c(2013:2015))
}


}

