% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename() generates a string}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the input year will be used to generate the file name, it is a string or integer}
}
\value{
a string "accident_year.csv.bz2", where year is the input year
}
\description{
This function use year as an input and return a string "accident_year.csv.bz2", where year is
the input year
}
\examples{
\dontrun{
   make_filename(2013)
}

}

